QA to Master (#3)
* updating poms for 1.0.2-SNAPSHOT development

* updating develop poms to master versions to avoid merge conflicts

* Updating develop poms back to pre merge state

* Moved Add Receipt Button to the top of the Page

* remove auto select drug name

* Shibwe pref (#2)

* Pharmacy Inventory app updates

* modified:   omod/src/main/webapp/pages/addReceiptsToGeneralStore.gsp

* removed date of manufacture and company name as requested t minimize un-nessesary field when taking in stock


* Added the required dependancies
