+++ api/src/main/resources/messages.properties
@@ -1,1 +1,1 @@
-${project.parent.artifactId}.title=Inventory App Module
+${project.parent.artifactId}.title=EHR Inventory App Module
