+++ omod/pom.xml
@@ -3,14 +3,14 @@
 
 	<parent>
 		<groupId>org.openmrs.module</groupId>
-		<artifactId>inventoryapp</artifactId>
-		<version>1.0.1</version>
+		<artifactId>ehrinventoryapp</artifactId>
+		<version>1.0.0-SNAPSHOT</version>
 	</parent>
 
-	<artifactId>inventoryapp-omod</artifactId>
+	<artifactId>ehrinventoryapp-omod</artifactId>
 	<packaging>jar</packaging>
-	<name>Inventory App Module OMOD</name>
-	<description>OMOD project for InventoryApp</description>
+	<name>EHR Inventory App Module OMOD</name>
+	<description>OMOD project for EHR Inventory Application</description>
 
 	<dependencies>
 
@@ -70,7 +70,7 @@
 
 			<groupId>org.openmrs.module</groupId>
 
-			<artifactId>inventory-api</artifactId>
+			<artifactId>ehrinventory-api</artifactId>
 
 		</dependency>
