+++ omod/src/main/java/org/openmrs/module/ehrinventoryapp/fragment/controller/AddReceiptsToStoreFragmentController.java
@@ -1,23 +1,27 @@
-package org.openmrs.module.inventoryapp.fragment.controller;
+package org.openmrs.module.ehrinventoryapp.fragment.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.apache.commons.lang.StringUtils;
+import com.fasterxml.jackson.databind.type.TypeFactory;
 import org.apache.commons.lang.math.NumberUtils;
 import org.json.JSONArray;
 import org.openmrs.Role;
 import org.openmrs.api.context.Context;
+import org.openmrs.module.ehrinventory.InventoryService;
+import org.openmrs.module.ehrinventory.model.InventoryStoreDrug;
+import org.openmrs.module.ehrinventoryapp.StoreSingleton;
+import org.openmrs.module.ehrinventoryapp.model.DrugInformation;
 import org.openmrs.module.hospitalcore.InventoryCommonService;
-import org.openmrs.module.hospitalcore.model.*;
+import org.openmrs.module.hospitalcore.model.InventoryDrug;
+import org.openmrs.module.hospitalcore.model.InventoryDrugCategory;
+import org.openmrs.module.hospitalcore.model.InventoryDrugFormulation;
+import org.openmrs.module.hospitalcore.model.InventoryStore;
+import org.openmrs.module.hospitalcore.model.InventoryStoreDrugTransaction;
+import org.openmrs.module.hospitalcore.model.InventoryStoreDrugTransactionDetail;
+import org.openmrs.module.hospitalcore.model.InventoryStoreRoleRelation;
 import org.openmrs.module.hospitalcore.util.ActionValue;
-import org.openmrs.module.inventory.InventoryService;
-import org.openmrs.module.inventory.model.InventoryStoreDrug;
-import org.openmrs.module.inventory.util.DateUtils;
-import org.openmrs.module.inventoryapp.StoreSingleton;
-import org.openmrs.module.inventoryapp.model.DrugInformation;
 import org.openmrs.ui.framework.SimpleObject;
 import org.openmrs.ui.framework.UiUtils;
 import org.springframework.web.bind.annotation.RequestParam;
-import com.fasterxml.jackson.databind.type.TypeFactory;
 
 import java.io.IOException;
 import java.math.BigDecimal;
@@ -29,7 +33,7 @@ import java.util.Date;
 import java.util.List;
 
 /**
- * Created by franqq on 3/21/16.
+ *
  */
 public class AddReceiptsToStoreFragmentController {
     private InventoryService inventoryService;
@@ -107,10 +111,10 @@ public class AddReceiptsToStoreFragmentController {
             String drugName = drugInformation.getDrugName();
             String unitPriceStr = drugInformation.getUnitPrice();
             String costToPatientStr = drugInformation.getCostToThePatient();
-            String companyName = drugInformation.getCompanyName();
+//            String companyName = drugInformation.getCompanyName();
             String batchNo = drugInformation.getBatchNo();
             String receiptFrom = drugInformation.getReceiptFrom();
-            String dateManufacture = drugInformation.getDateOfManufacture();
+//            String dateManufacture = drugInformation.getDateOfManufacture();
             String dateExpiry = drugInformation.getDateOfExpiry();
             String receiptDate = drugInformation.getReceiptDate();
             String institutionalCost = drugInformation.getInstitutionalCost();
@@ -134,15 +138,18 @@ public class AddReceiptsToStoreFragmentController {
             if (null != unitPriceStr && "" != unitPriceStr) {
                 unitPrice = NumberUtils.createBigDecimal(unitPriceStr);
             }
-            if(!StringUtils.isBlank(dateManufacture)){
-                DateFormat dateFormatter = new SimpleDateFormat("yyyy-MM-dd");
+            DateFormat dateFormatter = new SimpleDateFormat("yyyy-MM-dd");
+            /*if(!StringUtils.isBlank(dateManufacture)) {
 
                 Date dateManufac = dateFormatter.parse(dateManufacture);
+            }*/
+
+
                 Date dateExpi = dateFormatter.parse(dateExpiry);
-                if (dateManufac.after(dateExpi)) {
-                    errors.add("inventory.receiptDrug.manufacNeedLessThanExpiry");
-                }
-            }
+//                if (receiptDate.after(dateExpi)) {
+//                    errors.add("inventory.receiptDrug.receptDateAfterExpiryDate");
+//                }
+
 
             InventoryDrugFormulation formulationO = inventoryService.getDrugFormulationById(formulation);
             if (formulationO == null) {
@@ -162,7 +169,7 @@ public class AddReceiptsToStoreFragmentController {
             transactionDetail.setFormulation(inventoryService.getDrugFormulationById(formulation));
             transactionDetail.setVAT(VAT);
             transactionDetail.setBatchNo(batchNo);
-            transactionDetail.setCompanyName(companyName);
+//            transactionDetail.setCompanyName(companyName);
             transactionDetail.setCurrentQuantity(quantity);
             transactionDetail.setQuantity(quantity);
             transactionDetail.setUnitPrice(unitPrice);
@@ -173,7 +180,7 @@ public class AddReceiptsToStoreFragmentController {
 
             try {
                 transactionDetail.setDateExpiry(formatter.parse(dateExpiry+" 23:59:59"));
-                transactionDetail.setDateManufacture(formatter.parse(dateManufacture + " 23:59:59"));
+//                transactionDetail.setDateManufacture(formatter.parse(dateManufacture + " 23:59:59"));
                 transactionDetail.setReceiptDate(formatter.parse(receiptDate + " 23:59:59"));
             } catch (ParseException e) {
                 e.printStackTrace();
