+++ omod/src/main/java/org/openmrs/module/ehrinventoryapp/page/controller/SubStoreIssueAccountDrugPageController.java
@@ -1,10 +1,10 @@
-package org.openmrs.module.inventoryapp.page.controller;
+package org.openmrs.module.ehrinventoryapp.page.controller;
 
 import org.openmrs.api.context.Context;
 import org.openmrs.module.appui.UiSessionContext;
 import org.openmrs.module.hospitalcore.model.InventoryDrug;
 import org.openmrs.module.hospitalcore.model.InventoryDrugCategory;
-import org.openmrs.module.inventory.InventoryService;
+import org.openmrs.module.ehrinventory.InventoryService;
 import org.openmrs.module.referenceapplication.ReferenceApplicationWebConstants;
 import org.openmrs.ui.framework.SimpleObject;
 import org.openmrs.ui.framework.UiUtils;
@@ -15,10 +15,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 import java.util.Date;
 import java.util.List;
 
-/**
- * @author Dennys Henry
- *         Created on 06/06/2016.
- */
 public class SubStoreIssueAccountDrugPageController {
     public void get(@RequestParam(value = "categoryId", required = false) Integer categoryId,
                     PageModel model,
