+++ omod/src/main/resources/webModuleApplicationContext.xml
@@ -22,7 +22,7 @@
 
 	<!-- Add here beans related to the web context -->
 	<bean class="org.openmrs.ui.framework.StandardModuleUiConfiguration">
-		<property name="moduleId" value="inventoryapp"/>
+		<property name="moduleId" value="ehrinventoryapp"/>
 	</bean>
