+++ pom.xml
@@ -2,16 +2,16 @@
 	<modelVersion>4.0.0</modelVersion>
 
 	<groupId>org.openmrs.module</groupId>
-	<artifactId>inventoryapp</artifactId>
-	<version>1.0.1</version>
+	<artifactId>ehrinventoryapp</artifactId>
+	<version>1.0.0-SNAPSHOT</version>
 	<packaging>pom</packaging>
 	<name>Inventory App Module</name>
 	<description>Inventory App</description>
-	<url>https://wiki.openmrs.org/display/docs/Inventory+App+Module+Module</url>
+	<url>https://healthit.uonbi.ac.ke/</url>
 	
 	<developers>
 		<developer>
-			<name>francis</name>
+			<name>HealthIT</name>
 		</developer>
 	</developers>
 	
@@ -21,9 +21,10 @@
 	</organization>
 
 	<scm>
-		<connection>scm:svn:https://svn.openmrs.org/openmrs-modules/inventoryapp/trunk/</connection>
-		<developerConnection>scm:svn:https://svn.openmrs.org/openmrs-modules/inventoryapp/trunk/</developerConnection>
-		<url>https://svn.openmrs.org/openmrs-modules/inventoryapp/trunk/</url>
+		<connection>scm:git:https://github.com/uonafya/inventoryapp.git</connection>
+		<developerConnection>scm:git:https://github.com/uonafya/inventoryapp.git</developerConnection>
+		<url>https://github.com/uonafya/inventoryapp.git</url>
+		<tag>HEAD</tag>
 	</scm>
 
 	<modules>
@@ -32,7 +33,10 @@
 	</modules>
 	
 	<properties>
-		<openMRSVersion>1.9.8</openMRSVersion>
+		<openMRSVersion>1.12.0</openMRSVersion>
+		<kenyacoreVersion>2.4.1</kenyacoreVersion>
+		<kenyauiVersion>2.1.2-SNAPSHOT</kenyauiVersion>
+		<kenyaemrVersion>17.0.3</kenyaemrVersion>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 	</properties>
 
@@ -120,7 +124,7 @@
 												
 				<artifactId>hospitalcore-api</artifactId>
 																								
-				<version>1.5.3</version>
+				<version>1.0.0-SNAPSHOT</version>
 												
 				<scope>provided</scope>
 				
@@ -128,8 +132,20 @@
 
 			<dependency>
 				<groupId>org.openmrs.module</groupId>
-				<artifactId>inventory-api</artifactId>
-				<version>2.0.12</version>
+				<artifactId>ehrinventory-api</artifactId>
+				<version>1.0.0-SNAPSHOT</version>
+				<scope>provided</scope>
+			</dependency>
+			<dependency>
+				<groupId>org.openmrs.module</groupId>
+				<artifactId>kenyacore-api</artifactId>
+				<version>${kenyacoreVersion}</version>
+				<scope>provided</scope>
+			</dependency>
+			<dependency>
+				<groupId>org.openmrs.module</groupId>
+				<artifactId>kenyaui-api</artifactId>
+				<version>${kenyauiVersion}</version>
 				<scope>provided</scope>
 			</dependency>
